# Services/Server/song-content/openapi.yaml
openapi: 3.1.0
info:
  title: Mono Song Content API
  version: 1.0.0
servers:
  - url: https://ncm.zijiu522.cn
security:
  - appToken: []
paths:
  /api/public/song-content:
    get:
      summary: 获取统一歌曲、已发布内容、来源和生成状态
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/SongID'
        - $ref: '#/components/parameters/Locale'
      responses:
        '200': { description: 歌曲详情 }
        '304': { description: 内容版本未变化 }
        '409': { description: 歌曲身份待确认 }
        '429': { description: 请求过于频繁 }
  /api/public/song-content/ensure:
    post:
      summary: 幂等创建首次内容生成任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, song_id, locale, song]
              properties:
                platform: { type: string, enum: [NCM, QCM, QSM, KCM, AM] }
                song_id: { type: string }
                locale: { type: string }
                song:
                  type: object
                  required: [title, artists]
                  properties:
                    title: { type: string }
                    artists:
                      type: array
                      items:
                        type: object
                        required: [id, name]
                        properties:
                          id: { type: string }
                          name: { type: string }
                    album:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                    duration_ms: { type: integer }
                    cover_url: { type: string, format: uri }
                    isrc: { type: string }
                    platform_artist_id: { type: string }
                    platform_album_id: { type: string }
      responses:
        '200': { description: 已存在已发布内容 }
        '202': { description: 已创建或复用生成任务 }
  /api/public/song-content/jobs/{jobId}:
    get:
      summary: 获取安全的生成任务状态
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: 任务状态 }
        '404': { description: 任务不存在 }
...
components:
  securitySchemes:
    appToken:
      type: apiKey
      in: query
      name: token
    adminToken:
      type: apiKey
      in: header
      name: x-admin-token
  parameters:
    Platform:
      name: platform
      in: query
      required: true
      schema: { type: string, enum: [NCM, QCM, KCM, AM] }
    SongID:
      name: song_id
      in: query
      required: true
      schema: { type: string }
    Locale:
      name: locale
      in: query
      schema: { type: string, default: zh-CN }
...